requirements.txt
discord.py
psutil
python-dotenv
rich
databases[aiomysql]